Drop the idea with two char and sentence offset arrays.
The idea behind two arrays was, that all offsets focused on the first file and it could be interesting also to get the offsets focused of the second file. But the usecases are, that one JSON database will be given and the second file are only small JSON fragments.
=> A offset for the second file give no further infos.

And if a offset for the second file is really wanted; a simple swap of the CLI input file parameter is enough.