Create TEST for the intersection process.
With more specific words: Two TEST functions were added to check the number of tokens and sets in a full Exec_Intersection call.
For both tests the in the reposiory added file "intervention_10MB.txt" is used to gurantee, that the created number of tokens and sets are always equal.