Bugfix: The char offsets need an extra increment.
The reason is, that the original data uses the offsets with an increment for the spaces between the tokens. Without this commit the resuts are not in sync with the original offsets.
An example of the original data:
"tokens": [
"[",
"The",
"chemotherapy",
"of",
...

"abs_char_offsets": [
0,
2,
6,
19,
...

You see, that the offset of the second token is 2, and not 1!